fixed httpflood
httpflood works fully, and uses socket.

Long explanation: There isn't a way to spoof ip since the attacker must first complete the 3-way syn handshake before sending get or post, which can't be done if the syn-ack from the server is going to the spoofed ip.

Maybe can be used in a botnet? Don't do that though because its very illegal and not cool.